Keep preferred process name up-to-date
With this change, the preferred process name is fetched
once every second to keep it updated. This will prevent the
usage of an out-of-date preferred process name. A correct
and up-to-date preferred process name is key in the Task-Based
UX as it serves a signifcant role in the process list and in
startup profiling/tasks.

Without this change, the preferred process name is only set
on either (1) clicking a profile action button (which is not
always used to start a process) and (2) opening the tool
window (but the tool window can be opened before the gradle
sync has finished, resulting in a null name fetched). In
both cases, it is clear how one can end up with an out-of-date
preferred process name.

Bug: 277798461
Bug: 277798531
Test: updated
Change-Id: Iefd238573344c6e61036aab13a43df706233896c

GitOrigin-RevId: d205bf3f344146949fd0f8132e4c2a920a4b00dd