Add static preferred process entry
So that an offline device can be used to perform startup profiling
as well as an online device with no preferred process running.
This process entry is modeled as a DEAD process.

In the case the device does have the preferred process running, the
static preferred process entry wraps the real corresponidng process
and displays the pid and manifest configuraiton, along with the
notable android head icon as a visual indicator that this is the
preferred process.

Included and heavily tied to performing this change is a refactor
to how the home tab selections are made. Now, all tasks are enabled
and based on the process and task selection, the startup profiling
dropdown and start profiler task button state is adjusted (change
in dropdown options and enablement respectively).

Fixes: 277798531
Bug: 277798531
Bug: 277798461
Test: updated
Change-Id: I5884b37a53e7218dcc48c421cd6ca34091945179

GitOrigin-RevId: b55c42370941b058dd443c392cf191d3c886a964