Fix slow previewElements being called on the ui thread
The call to previewElements() was being made on the ui thread when it
should have been done on the worker thread. Furthermore, the preview
elements which were supposed to be previewed was already computed and
passed in the file as the filePreviewElements parameter, which was
ignored.

This change uses the filePreviewElements parameter, whose value is
calculated on the background thread, to fix slow operations on the ui
thread.

Bug: 227474520
Test: manual
Change-Id: Iabbeb13c53487a94fd12930b5bf02ff9310966cb

GitOrigin-RevId: 053fc013d67b6b39286938ac7dca08d0eaf89f6c