Cleanups from previous APK / Bundle action changes
Prefer to use "App Bundle" rather than just a bare "Bundle" to be more
consistent with developer documentation.  Also adjust some class names
for consistency.

Adjust the Export Signed Package wizard to have the choice-oriented
flow be active always, as it will be given its single use.  Remove the
now useless check for whether the choice-oriented flow should be
shown, which was reduced to a simple project nullability check.

Bug: n/a
Test: existing
Change-Id: Ie4dcc7e5787bd6aa75fe54853cc268abd29900ca

GitOrigin-RevId: 3bcda16f88f9efdb4f788160ff842af9e6f23715