[AUA] Make sure update to jdk is in jdk table
... and modify jdk of LATEST to 17.

When AUA tests are run, it is expected that the upgraded test projects
end up with the JDK used in LATEST. This TestAGPVersion is using the
default JDK, which is the one used to run the test, instead of
explicitly saying what version should be used. Since the tests only have
the `from` state JDK in the JDK table, tests that upgrade the JDK
version fail since ProjectJdkRefactoringProcessor will be blocked.

This change fixes two issues:
 - Make sure that the JDK required by the upgrade AGP is present
 - Explicitly say that LATEST requires JDK 17

Bug: 311745645
Test: Ran AUA targets using --config=jbr-next
Change-Id: Ib62297786940e33228d42c325c0308664fad9dd4

GitOrigin-RevId: a84ba17b93b576d05b8a172db152e36f19249e9f