Fix error in TomlErrorParser
The problem was that TomlErrorParser did not consume exception
stacktrace from the output, leaving this for other parsers to read.
JavaOutputParser in turn looks for 'Caused By:' strings producing error
messages from them.

As this is a 'Build failed with exception' parser, we can consume all
output till the end, as it is done in GradleBuildScriptErrorParser.

Fixes: 326059775
Bug: 306405462
Test: added TomlErrorTest
Change-Id: I929a19603ef31cc4466ef678853b3eb149dccc42

GitOrigin-RevId: ebbf31bdab1db14e0edbcfea24657bf477249995