Make "play store compatible" toggle the default for wear
... devices.

There are still a lot of play store screenshots that fail the screenshot
requirements. According to the data, users use the "Display Shape"
screenshot option a lot more than the "play store compatible" one.

It's impossible to know if those screenshots are the ones that are used
on the playstore. However, "Display Shape" was the default toggle option
which could explain the higher usage. This CL changes the default to
"play store compatible". This is aimed at reducing play store rejections
due to incompatible screenshots.

Bug: 235061954
Test: updated ScreenshotViewerTest
Change-Id: I2ab75d252f1cc1998e6fb856ec8ac4b60fe12859

GitOrigin-RevId: 12e467af46240541c84b5293187101648dc1f47f