Use layout binding cache for ignoreDumbMode
The error log involved in this change is meant to help identify any dumb
mode accesses remaining of LayoutBindingModuleCache, so that we can
remove them. However, it turns out that the latest platform updates
makes it a valid case *sometimes* for indexes to be read in dumb mode.

This change updates the conditions under which we log an error;
hopefully the error will go away following this change and we can then
remove the check entirely.

There is no test for this because trying to effectively hit this
requires accessing the cache in dumb mode; but about 1/100 runs this
hits an idempotency check from the platform that results in a recursion
error. Given that the DumbModeAccessType functionality is experimental,
it's not surprising that there are some rough edges, especially around
testing. Since the ultimate goal here is to remove these conditions
(which would just let any potential future errors surface via
exceptions), I'm not adding a specific test for this case.

Bug: 322209412
Test: n/a
Change-Id: I725e83d8c1561ba7f5a4d3f8d7bfb45556f8de28

GitOrigin-RevId: 0ffd5b37becb914f3b62d92a22c0a9d32673cbba