Refactoring to SyncFailureUsageReporter
To provide full picture of the failure we started to attach build
output parser generated errors to the failure details reported here.
This introduced flakiness in tests and reporting itself as by the
time we send event BuildOutputParsers might not have process all
the output yet.

To fix this we need to wait for FinishBuildEvent to guarantee all
events were received and only then report the collected stats.

Bug: 324894199
Bug: 306638551
Bug: 306405462
Test: Adjusted SyncFailureUsageReporterTest to the new flow
Change-Id: I33f99c39b136b22d70ea7d39b5c04cbe0aabea17

GitOrigin-RevId: ef2d1143ffc087b6cacd7ddcbffe322a40392f0f