Check for context sharing before calling StudioBot#sendQuery
We can't call StudioBot#sendQuery if the user hasn't opted-in to share
context using the Studio Bot onboarding. By calling the API regardless,
we will throw an exception that will not be handled properly. This
change modifies this flow so we properly handle the exception and show
a relevant message to the user.

Bug: 322759144
Test: NShotXmlToComposeConverterTest
Change-Id: Id60daf1fadbf27c5bcc9c8dcb8ec058317e7aee0

GitOrigin-RevId: 2ae70996cc49f06712a02001e0bf60f303042bd4