Add repository library search by module
If user does not type `:` in library search(PSD),
we are not sure if we should look by name or group and
whether user is aware advanced syntax with asterisk.
In this case we'll add asterisk at the beginning and
end of search string and search in "groupId:name"
strings.

Test: GoogleRepositoryBaseTest, MavenCentralRepositoryTest
Fixes: 133176003
Change-Id: I352603a527bead165a92a76c383869d265a0fece

GitOrigin-RevId: 25441e186128c9f1140a4f2ab9d0c667a4392888