prompt builder: Provide a way to append to existing prompts
This CL adds an `appendToPrompt` method that allows for adding
to an existing prompt. In order to do this, this CL also relaxes
the restriction that all prompts have to end with a user message.
This is useful in scenarios where one part of the code may
add a model response to an existing prompt, and then some other
part may follow up with the existing prompt and append a new
query from the user before sending it to the model.

Bug: N/A
Test: SafePromptBuilderTest
Change-Id: Ia1a4375ed7f90578aa5677ae8e975679d305fa8a

GitOrigin-RevId: 24534dbfba12f180ea3adc88b9570488c88fb995