Remove unreachable task icon disablement code
In a previous UX implementation, task icons could be
in a disabled state where they had a special disabled
icon and were not selectable. Now, all tasks are always
selectable and use their respective task icon.

The isTaskEnabled boolean param was hard coded as true
in a recent change, so all of the disabled icon code was
unreachable anyways. Thus, this change removes all the
unreachable code.

Bug: 277798754
Test: existing
Change-Id: Ib64466c48c406b32ad54fdc196a259199396b392

GitOrigin-RevId: a6872a96d52584d8dbe999e66b033771f95665e3