Reuse Android SDKs for testing
With the 2024.1 platform, there's an assertion to prevent adding the
same SDK multiple times to the Project SDK table. To fix this, we
instead check to see if an SDK exists for a given Android platform and
use that for configuring the module before creating a new one.

Tests: existing
Fixes: 326288681
Change-Id: I65458107d71a3600c468689f35bf344d0a47203f

GitOrigin-RevId: 53ceb34dc9da70597d0c4ba834c1f2de8c9a3454