Try to fallback JDK to one required by Gradle
... when there are issues with the project JDK.

We currently try to fallback to the default and embedded JDK when there
are issues with the project JDK, regardless of what version of Java is
required by the project. With this change, we now try to use a JDK from
that is compatible with the used version of Gradle. If there is no JDK
that can be used, now we show an error notification saying what version
is required by Gradle and pointing the user to pick one compatible
(opening the Settings, where they can pick one detected or download a
new one).

In the case the Gradle version cannot be detected the behavior remains
unchanged (try default and then embedded).

Bug: 313677054
Test: GradleSyncInvalidJdkConfigurationIntegrationTest
Change-Id: I170ee201d9a115bbd3ba6fef5a81f752bcc3b091

GitOrigin-RevId: 0393377e8fc0f6d4b7135eb7881e7545c6d4f84f