Remove profiling warning in live task
In the live view, a ‘Profiling with complete data…’
or "Profiling with low overhead..." warning is shown
at the top. This has been removed because any warning
regarding the build type (profileable or debuggable)
will already be shown before the live view starts
(next to start profiler task button in the home tab).
Thus, the warning message has been removed.

Note: It was removed and not flagged away because the usage
of the LiveStageView is behind a flag.

Fixes: 328560283
Test: updated
Change-Id: Ia3beb765c13509508ede5750f87f766dbee637fa

GitOrigin-RevId: adece16d791957a22772f64cde33f5cd72978999