Check if skins exist before copying
The DeviceSkinUpdater attempts to copy skins from the Studio binary into
the SDK folder. However, in the case of a system image that comes with
its own skin, they will not be present in Studio, so the copy will fail.

Note that we only reach this point because the SystemImageDescription is
null. The DeviceSkinUpdater is invoked in various places. One place is
upon the initial device profile selection, before the system image is
selected.

Additionally, the FileUtils.copyDirectory() method throws an
IllegalArgumentException, which is not caught by the surrounding catch.
Thus, this exception propagates to the top level, and the selection of
the device does not take effect: the UI shows the just-clicked device as
selected, but the underlying model still holds the previously-selected
device.

We need to check that the directory actually exists before attempting to
copy it, and if not, return the "skin" input, as documented in
DeviceSkinUpdater.updateSkin.

Bug: 326253336
Test: DeviceSkinUpdaterTest
Change-Id: I87e70dd54b15241426f1d1d39a47255d292e28e3

GitOrigin-RevId: 84344c545639a5277c01c088d53b0d7ba172d30b