Remove lock from LightBindingClass creation
This is the last remaining lock usage in LayoutBindingModuleCache, so
the source of some deadlocks is now completely removed.

BindingLayoutGroup is used to store the light classes and ensure there's
only one list of classes per group. That was done with a generic
mechanism (UserDataHolder) which unfortunately required locking, since
the semantics of that interface don't play well with concurrency.

The structure resulting here is a tad strange; the classes are created
by LayoutBindingModuleCache, but stored in an AtomicReference on
BindingLayoutGroup. But I think the result is fairly clear in terms of
implementation, and is safer code that doesn't require locks.

Bug: n/a
Test: Covered by existing
Change-Id: Ib3884bd469b9b96cea12c1905f9a663a158824c5

GitOrigin-RevId: 8ac0487e597e0d72ac2342bb221e68f7cb9b365e