Add Google maven repo to settings file
When the repository block is declared in dependencyResolutionManagement
in settings file, we add the quick fix link to add google repo here
instead of individual build files. Only if the repository block is not
declared in settings, the quick fix add google repo in module build
files. In addition, we also check if the google repo is declared in
settings file before we add a quick fix link to add google repo.

Bug: 327193928
Test: UnresolvedDependenciesReporterIntegrationTest, AddGoogleMavenRepositoryHyperlinkTest
Change-Id: I5dd25842d10ab88fc532065273b5269e000d0a70

GitOrigin-RevId: 268f11ef436c1e8d1739c9bde3ddf2220d6c39f9