Make a couple of tests slightly more robust
In AddGoogleMavenRepositoryHyperlink, be a little more careful to
save and retrieve the Dsl models after making changes, to be sure that
we have a view on the correct data.

In UnresolvedDependenciesReporterIntegrationTest, the need for an
apparently-spurious settingsModel to apply no changes to was because
of disagreements between the cached virtual file layer, Psi, and the
IO files.  Modify the settings file before opening the prepared
project, making sure to inform the VFS layer of the changes.

Bug: 327193928
Test: existing
Change-Id: I9d297335665edc06e373053fa4b781166e3f6b32

GitOrigin-RevId: 7a30b0713f3a61fb7d968f8ab6c34dbea7f14d97