Change signature of previewElementFlowForFile to take a
... project and a PreviewElementProvider as parameters and rename the
function.

The function now returns preview elements from the provider anytime a
file changes in the project, instead of searching from a
FilePreviewElementFinder specifically.

The reason behind the change is to abstract away where the preview
elements are coming from. In CommonPreviewRepresentation, we only deal
with PreviewElementProviders and don't interact with any
FilePreviewElementFinder directly.

This change is a step in making this function usable by
CommonPreviewRepresentation without having to make
CommonPreviewRepresentation use a FilePreviewElementFinder.

Bug: 328758600
Bug: 326412365
Test: updated PreviewElementFlow and PreviewElementFlowTest
Change-Id: I9afd1f1ead8927961026f9c34cf80c5a5ae48ac0

GitOrigin-RevId: da4a39c7ad78305cd6f6767e680bfcf4fee503eb