Evaluate light binding class config immediately
LightBindingClassConfig is used to populate methods/fields/etc on
LightBindingClass instances. However, this is currently done in a lazy
manner, and recent bugs indicate it's likely that some of this is
happening during indexing. If BindingXmlIndex is unavailable, then the
returned data would be incomplete or missing, and bad data could end up
being cached.

This change attempts to address the above by eagerly calculating the
config's values when the config objects are constructed.
(LightBindingClass retains lazy logic for constructing its PsiFields
and PsiMethods, but it is based off of this config that will be more
stable.)

I'm putting this change behind a flag in case there are unexpected
performance issues that crop up. I don't expect that to happen, as
evaluation of the config is relatively lightweight; but it does involve
accessing various pieces of data from the platform, so I'm erring on the
side of caution.

Bug: 330744400
Test: Included
Change-Id: Ibf230e17e65a3056d1e97e4a257e46a2e3748a90

GitOrigin-RevId: 07c82c571598d5a244e01360e0707f9ab4a77ce8