Request sync if unable to link module group
In some scenarios it seems like workspace caches
don't load correctly. More specifically, content
roots for some modules are missing, which causes
the IDE to remove those modules. Their data nodes
remain available, but during "linkAndroidModuleGroup"
not all data is available.

Previously, this caused crash reports to be generated.
However, given this is an issue in the workspace
model (that is also very hard to reproduce), a more
suitable behavior is to request sync if any of data
is missing (in line with how the rest of missing data
is handled).

This change also tries to unify kmp and non-kmp module
linking.

Bug: 300771593
Test: this issue is very hard to reproduce, and
more details are provided in the issue comments

Change-Id: I3fe44b36487e53d3575e1fc9683b6aba2975ab41

GitOrigin-RevId: bc74778f20ab70973098d6b39c0fe06f5d07b058