[lint,SMTH] Allow lint to consider Something files
Add the language type to the lint plugin, and follow-through in the
IdeSupport classes.  Make the tests of the driver and annotator
conditional on the flag, which doesn't affect the underlying
machinery (so the added test doesn't need an override) but does affect
lint in the IDE, both live annotations and code analysis in the
Problems window.

Bug: 330888875
Test: manual, added Gradle visitor test
Change-Id: I05a9e04e85f2dcc6d733e60ee488959778f75a17

GitOrigin-RevId: a654683252a40aa32bb6324e0cb4341d4054237e