[SMTH] Implement a UAST plugin for Something
This isn't a complete implementation, though it's also not broken: all
interfaces are implemented with no error-throwing stubs, but it's
incomplete in that some classes methods are less helpful than might be
desired (in particular, SomethingValues that aren't literals are
poorly represented at present).

This is enough to get Lint running in the IDE on
`build.gradle.something` files.

Bug: 330888875
Test: manual, integration test in next CL
Change-Id: Ie8879a55a3955c55331d8d95b10e2e16e8ad2f5f

GitOrigin-RevId: 48c6fd0d62d21be8ebad0b9c359ead155fcc2ac3