Add test validating namespace update
b/74451966 calls out that resource repositories do not correctly update
when namespaces change. This is still technically true; a resource
repository constructed before such a change will not reflect the change.

However, since the bug was filed there have been a number of changes to
the resource repository system. Most importantly for this issue,
consumers no longer should be storing references to repositories;
instead, the fetch them on demand when they were needed. That means that
only a new repository fetched after the namespace change need return the
new namespace.

This behavior is now working as expected, so I'm adding a test to codify
it.

Bug: 74451966
Test: test code
Change-Id: I034d1611a977e19ec7b1fc4d54d995b0eaf4cf57

GitOrigin-RevId: a80449277a15e395c1f2accd6d29abf2d0194b82