K2: Create KtCodeFragment for new PSI in ComposeModifierCompletionContributor
In order to use analysis API (AA) on the newly created PSI, we have to
create `KtCodeFragment` for the correct session. Otherwise, AA will
cause exception because the dangling module session is not for code
fragment.

This commit updates ComposeModifierCompletionContributor to create
`KtCodeFragment` for new PSI used by the contributor, which fixes the
exception caused by the wrong dangling module file.

Bug: 330760992
Test: ComposeModifierCompletionContributorTest
Change-Id: Iea43948663f7189def0e447416f0e51afaf3771a

GitOrigin-RevId: f889a34a059e7730dd6954ba37d91fdd91732977