Handle different editor themes
The editor can have a different theme from the rest of the UI. Since we
use an editor to display the stack trace, the editor theme is used
there. However, in the normal case, we change the background color of
the editor to be the same as the background color of the panel, which is
a problem if the themes are different.
Changing to only update the background if the editor theme selected is
the default for the main theme.

Test: none
Fixes: 327390563
Change-Id: I6edf79e3ff0c0f1cb5d76604d486bffeddb55fda

GitOrigin-RevId: 271c24721bd3cf6a2bdd8f88e3a63bbd27785299