Add utility to check for ScreenshotTest file
There are some utilities that check if a file is
a test file and this is being extended to also check
if the file is a screenshotTest file.

The main usage of these utilities is currently compose
previews where we make sure we are not running Previews
on a test file.

For ScreenshotTests, this is the same, as in we do not
run previews themeselves, but rather tests on previews.

Bug: 302523774
Test: N/A
Change-Id: I683c65bcbf5e65df6efad365b844b94fd990de87

GitOrigin-RevId: 68c7740409849b3bece96cad57df99dd659ede0e