Track the latest ADB in AndroidDevice
AndroidDevice instances are rather ephemeral, so ordinarily when one of
the inputs changes, we just create a new one. However, one frequent
operation on AndroidDevice is looking up an IDevice using its
DdmlibDeviceLookup after launching it. Since booting an emulator can
take a little while, it's not too uncommon that ADB might restart while
this is happening. When this occurs, the DdmlibDeviceLookup holds a
stale reference to the old AndroidDebugBridge and fails to find the
device.

Instead, we can just always refer to the latest AndroidDebugBridge
instance when looking up devices.

Also, increase the timeout to 10 seconds when looking up the device, so
that a busy system doesn't trigger a timeout.

Fixes: 322375280
Test: manual
Change-Id: I61ce39f30554424d38558cb9bdfca765ff6aeaa5

GitOrigin-RevId: 4f6aa99705c0c59819a35e914438563e5f45011f