[SMTH] SomethingProperty to version catalogs entity references
Although these references work (to the extent that they are being
tested), in practice the user experience is not yet optimal for
subgrouped entries: I believe because of the difference between the
names of the elements (e.g. a name of "foo" in the context of a
reference of `libs.foo.bar`, corresponding to a key segment named
"foo-bar").

Also it's not actually clear that this will be the eventual syntax for
version catalog references, although the design goal of being able to
directly use declarative Gradle files as Kotlin Script build files
would suggest that this at least stands a chance.

At least some of the usual affordances work.  The `Rename...` action,
for example, works for non-subgrouped elements.

Adding the dependency from the project-system-gradle tests to the
something psi support also changes the associated icon type in the
Android Project View (for tests; there is no change in production).

Bug: 328257084
Test: manual, added
Change-Id: I5c0d7e8c8a1b594f09eaf9c0be8209da4bb85229

GitOrigin-RevId: f70556d00dfd554b5a780df41b82e37fbb55fb82