Clear Panel When Selecting a Device That is Offline
There's an edge case where an online device has a lot of backlog, and we
switch quickly to the online device and then to an offline device.

Since updating the panel is asynchronous and contains multiple thread
switching, we need to prevent message from getting added from the online
device after we already switched to the offline device.

This is done by passing the selected item as a context which is then
passed back and compared to the currently selected device.

Fixes: 275034206
Test: Manually
Change-Id: I4a6913407be36cfb687fff43828bcb556ff5dcdb

GitOrigin-RevId: 1fa02f256b5aa622cc6aa4b17017c5a2d10b8616