Separate deprecation presentation changes
This change separates out the functionality related to changing the
presentation of deprecated items in different contexts from the more
generally named "AndroidJavaCompletionContributor".

This change also enables this contributor for Kotlin code, even though
we don't currently cross out deprecated items in Kotlin. We do this
so that if that feature is added for Kotlin, we can catch cases where
we would be showing things crossed out, even though they might not be
deprecated _in that specific context_.

Bug: 293306513
Test: AndroidDeprecationPresentationCompletionContributorTest.kt
Change-Id: I17f28a469bce8f53da868845cb67768bc5f67e36

GitOrigin-RevId: 197610a515bcf46dc078f970118f81fa1c4beadb