Fix build KMP project for non-android source sets
In Studio, the Gradle list of tasks isn't populated by default
for performance reasons. This means that when building the entire
project, each module is evaluated individually and build tasks are
assigned to compile the source/test code, happening this logic
on GradleTaskFinderWorker. However, in the case of KMP, the non-android
source sets are considered as java modules, which isn't exactly
correct since those do not necessarily contain the expected java gradle
tasks like "testClasses" in case for example of ios source set.

Since Studio has to "guess" the tasks name since there aren't
available, for all non-android KMP source sets just "assemble" task
will be executed when building the specific module delegating in this
way the responsibility to who is registering the Gradle task. However,
since "testAssemble" lifecycle task doesn't exist, the tests will not
be compiled and users have to do it separately.

Bug: 292110316
Test: testKotlinMultiplatform_ios + existing
Change-Id: If85c943bea7f39e6921c08ec0ee6153cc47eafff

GitOrigin-RevId: 15e192d94441cbb9d41f7a431fab1e65cfc3556f