Introduction of StudioIconsCompose
StudioIconsCompose is a generated reference file of SVG icon painters
that provides convenient access to all icons unique to Android Studio,
specifically to be used for Compose UIs. The set of icons is based on
the resources in the artwork module, which is regularly updated by the
UX team based on the team icons source of truth in Figma.

For Swing UIs, StudioIcons.java should continue to be used.

Test: Visual inspection of local build
Bug: n/a
Change-Id: I3dc3f05c55c3f9c003911e7e4aed8c7f7e58d15b

GitOrigin-RevId: 33a0c8939a7f0e596641d6ff01d27cfa219f6fb9