Fix Prompt models after yesterday's CL
Yesterday's CL broke blobs handling for the multimodal code with changes
in the models. This CL, in conjunction with ag/26747052, fixes all
newly introduced regressions.

In particular, this CL adds the missing equals/hashCode to BlobChunk: as
suggested by the IDE inspection, not doing it can cause bugs.

ByteArrays' equals/hashCode don't check the array contents but only
their references, so identical instances would not be equals.

This CL also avoids calculating base64 representations of blob data
until needed, and avoids including it in equals/hashCode since it's
directly derived from the non-base64 data.

Then, the non-base64 data is now accessible externally — it's needed
to convert to/from other blob models.

The Model interface was made to expose a set of supported blob types
instead of a list, which is more correct and makes looking up faster.

Bug: n/a
Test: not in this module, but covered in aiplugin/core
Change-Id: I398c82b8e29003c7d49a18025ffd6ed9592180bc

GitOrigin-RevId: 59be9944f67c411a549802b1297eb32bf7a75002