Don't drop AAR class caches if nothing was cached
JetBrains has made the following change in the 2024.1 update:
https://github.com/JetBrains/android/commit/421e67e5b69afbea527df5b88f59429c6e662112

The underlying fix (only dropping PSI caches if they contain something
that needs to be flushed) makes sense, but our code has changed so it's
not a clean cherry-pick. This change makes a similar fix, but with the
code that exists after ag/26065707.

Bug: 332578858
Test: Covered by existing tests
Change-Id: Ib9fdd0bb8202c92a6369b43fd1e61d0c1d42fed0

GitOrigin-RevId: 9f9401c7d883d058e51fee17df0f9b8faa3124ac