Auto open recording on import
...to simplify the user flow. Also, this keeps the profiler's
on-import behavior consistent from the Session-Based UX. As in
the Session-Based UX, on importing a trace, it was automatically
opened. With this change, in the Task-Based UX, instead of just
adding an entry to the recording list on import, it also is auto
opened. Thus, in the Task-Based UX, opening a trace via Macrobenchmark
now works as it did previously (auto opens the trace in the Profiler
on selection).

Fixes: 330452930
Test: none
Change-Id: Ia14cff31cfa8716834ae0d80182b02d8ffd4414e

GitOrigin-RevId: 6cdda71348a5539ab5abc072b49654ec68ef507c