Ensure task is disposed
Apparently the dumb mode task queue only partially takes ownership of
its tasks, in terms of disposing them. It will dispose them after they
have been run - but it never actually registers itself as a parent/owner
of the disposable task. This can lead to the task appearing to leak
(even though the queue will dispose the task when it is disposed
itself).

The queue does, however, have logic to ensure that it won't run the task
if it's been disposed externally. So we can work around this by
registering our own disposable parent. In most cases this won't make any
difference; but it ensures the task has an appropriate parent and won't
ever get lost.

Test run x500:
https://fusion2.corp.google.com/invocations/3d1ab9a7-991d-4029-a31a-e5d90b4f0860

Fixes: 331293193
Test: Covered by tests that were failing
Change-Id: I49f085795ce625cbe4c7d2d9ad662ef5796df764

GitOrigin-RevId: 65e0d9f8b60dcc94e9cf28737e33a341bbf7e23a