Ensure lint checks are run on opening a layout file
For layout files, the issue panel is updated with lint errors after
a rendered has completed. But on opening the file, the render can
complete before the render listener to run lint is set up. When that
happens, the lint issues are not displayed in the issue panel until
the preview is rendered again.
This fixes the issue by running the lint checks after adding the
listener.

Bug: 332652415
Test: test added
Change-Id: Id9f53db0c57c74f7d505be285175666464e9c69a

GitOrigin-RevId: a3539ff967988f51dc91d8725da30b8576076d5b