Wait Until Indexes Are Ready
ContentEntryUtilTest#createContentEntry() modifies the Module model,
which can trigger Indexing, which itself can do a lot of things.  If
we don't wait for indexing to finish, we can end up racing with the
indexing, which among other things wants to instantiate the
TemplateDataLanguageMapping service.  Meanwhile, our test
infrastructure wants to clean up the TemplateDataLangaugeMapping
service, if it has been instantiated.

This can end up racing in serviceIfCreated called from
TestApplicationManager#tearDownProjectAndApp: we can attempt to find
the service and fail (because indexing has not created it yet), then
fall foul of the check that the service being requested is not a
Component (because indexing has created it in the meantime).  This
explains the somewhat confusing error message in the associated bug.

Fixes: 331810728
Test: existing (ran 1000x; previously failed in ~50 iterations)
Change-Id: I91f9e661d225f6aeb4eedc54cde0925f69025440

GitOrigin-RevId: 7679cad52e96eeaad7657a6fa8f151c4a8cb961f