[Dsl,.kts] Fix abiFilters in per-variant `ndk` block
The interface class (which is what is visible to the KotlinScript
compiler) does not define `abiFilter`/`abiFilters` helper methods,
only exposing the `abiFilter` val as a mutable set.  Describe the
model property appropriately, and adjust the Dsl property description
maps to reflect the available operators more correctly.

In doing so, we can now use our predefined AUGMENT_LIST description,
which allows the removal of an extending `addParsedElement` method.

Bug: none filed
Test: existing tests adjusted
Change-Id: If41265d3f4b8412e3ca6274cb7cdb59b0055c84d

GitOrigin-RevId: 9d114bc2e32fd9f1b8c5900e83bac39053f51c5f