Remove special casing in java dsl block
This was needed for historical reasons as in groovy defining java
plugin configuration on the top level was preferable at some point.
But:
1) it complicates adding toolchain support to java block;
2) top level java configuration is deprecated (replaced with java block
since Gradle 4.1) and is scheduled for removal in Gradle 9.0.

With this change we will start writing language level to java block in
groovy, same as in kts.

Bug: 333693199
Test: updated JavaModelTest
Change-Id: I2de4afe3554255d81dee8061ec90e4bbd4b7bff5

GitOrigin-RevId: 21e373d65d05e8da63b135a8a04d3b7827fab7f9