[SMTH] Support block comments
Grammar-Kit's token definition mini-language does not allow the
definition of alternate lexer states -- or if it does, it's completely
undocumented -- which means we need to drop down to the .flex lexer
definition "by hand".  Implement the IN_BLOCK_COMMENT lexer state,
returning BLOCK_COMMENT_CONTENTS tokens and tracking the comment
nesting state.  (At the moment, we do not attempt to distinguish
between simple block comments and KDoc-like comments).

The fact that we have to define our own lexer state means that we can
no longer simply re-run the lexer generator on the something.bnf
Grammar-Kit file.  I think that's OK; the flex definition is likely to
remain fairly clear.

Bug: 328257084
Test: added lexer tests
Change-Id: I31ec26dcacb0dcd8770c8de0a942a78e7cffd9da

GitOrigin-RevId: b414788a6e6383f647b84aaece34de048f358f1c