Ensure that Preview Elements are refreshed by Fast Preview
When Fast Preview finishes compiling, it will update the
ModuleClassLoaderOverlay. Preview Paramater classes might have changed
in that compilation so we need to ensure that they are updated.
This change introduces a new NotificationManager that can be used to
detect any changes in any ModuleClassLoaderOverlay so the elements can
be re-instantiated.

Fixes: 328229333
Test: NotificationManagerTest, ModuleClassLoaderOverlaysTest
Change-Id: Id41e179a150e61ceb6ac39367c7a2c51a56f1b49

GitOrigin-RevId: 8cc12f6c7088beb7ebcc564a995e17da597a0d05