[SMTH] Implement a BlockGroup PsiElement in the parser
The BlockGroup holds what in Kotlin would be the action lambda: the
braces and all the entries.  This is needed so that the UAST language
plugin for Something can refer to the BlockGroup as its own sourcePsi,
mostly for the purposes of test infrastructure (as otherwise the Java
plugin inserts carriage returns when generating a render string when
running on Windows).

This requires some minor changes in the Dsl parser and writer, and in
the code style settings for the Something language.

Bug: 330888875
Test: existing tests adjusted
Change-Id: Ieda21e56f3bd44fff8f093b84fb1e5f0d71ff0c3

GitOrigin-RevId: 30b1db5208c71ba6f2b88bf529af740768d6b1a4