Convert AnnotateQuickFix and AddTargetApiQuickFix to ModCommand API
This change introduces ModCommandLintQuickFix as an alternative base
class for Lint quick fixes, switches AnnotateQuickFix and
AddTargetApiQuickFix to use it, and updates various test fixtures to
work with both kinds of quick fixes.

I chose to start with these two because their conversion unblocks
the removal of one K2-resolve-on-EDT exemption. And they are good
examples of two different kinds of fixes. Some of the declarative
fixes based on LintFix can be converted to a combination of
corresponding ModCommands (ModCreateFile, ModUpdateFileText, etc.)
and some manual bookkeeping of which elements to make writable.
PSI quick fixes seem to have a simple(r) migration path via
PsiUpdateModCommandAction, although we need to remove all
prepareElementForWrite() calls and replace isPhysical() checks.

Note: Lint quick fixes that launch actions, navigate to URLs in the
browser etc. are not convertible to ModCommands, and will probably
remain DefaultLintQuickFixes forever.

Bug: 319287252
Test: existing
Change-Id: I96ef6e8340bb788b0d1ab0e055cd5ca856bed105

GitOrigin-RevId: 244e40c438d3cb20492b407bdb9da4317e881d32