Allow fractional scale greater than one
This change enables fractional scaling of the device display image when the available space in the Running Devices window is larger than the device display. By default, the scale of the display image is an integer number to avoid scaling artifacts, but the Zoom to Fit action remains enabled allowing the display image to be scaled up to occupy the whole available space.

Test: EmulatorToolWindowPanelTest
Bug: 332950382
Change-Id: I703a51935b6747c7a3ca92bbcf446a9beef381eb

GitOrigin-RevId: 46c45fea00ee51ac68019db28d75aa45973e01f0