Add java support to DeviceSpecInjector
String concatenation in the device annotation parameter does not work
properly in Java. This is mainly due to the JavaLanguageInjector that
takes precedence over other language injectors. The JavaLanguageInjector
does not handle string concatenation with references to other variables
properly.

This change also renames DeviceSpecInjectionPerformer to
KtDeviceSpecInjectionPerformer to properly convey that it only supports
Kotlin at the moment.

Bug: 333377591
Bug: 228294269
Test: updated DeviceSpecInjectorTest
Change-Id: I7a87b536c622e4092a438b513fc2f7eb834c67a3

GitOrigin-RevId: d35fde29a107d7423d0aa5826382774a0e867c50