Cleanup SyncAnalyzerManagerImpl.kt
1) Retrieve application service properly
2) Remove unnecessary code around feature flag. Flag check was moved by
Jetbrains to outer level which makes other checks obsolete. It was
previously checked on later stage to allow metrics gather and only block
UI. But as we do not have plans to turn off the flag this does not
matter for us anymore.
3) rework disposal order slightly. The problem was that previous fix
of disposing on project in LongDownloadsNotifier results in task future
being attached to project instance for project's life which is not good.
Instead in the manager add project disposable task to trigger cleanup
of this task related data. This is needed only in case buildFinished
disposal flow is not triggered properly on project close and makes sure
project related information is removed from this application service.

Bug: N/A
Test: existing tests
Change-Id: I6218751fe9f8ba46bc0f2cd0860973ff52c103f3

GitOrigin-RevId: 2d641e1ae0ef15ce1c60843d56faa37cf0602f20