[android] Create Quick Definition views in background
With the 2024.1 merge, the construction of
`PsiImplementationViewElement` asserts that it's on a background thread.
Since our quick definition is currently creating it on the event thread,
we need to put that behind a progress indicator.

This is corollary to how the platform handles the exact same case for
their standard `ImplementationViewSession`; see
PsiImplementationViewSession at
https://github.com/JetBrains/intellij-community/blob/master/platform/lang-impl/src/com/intellij/codeInsight/hint/PsiImplementationViewSession.java#L79

GitOrigin-RevId: 2994af9eaaf327d37ce61edfe805b4e2190dc4bd