Coalesce provider tests harder
Conceptually, the ProviderTest scenarios have two parts: one is the
setup of the project, variant, device, and so on; the other is whether
the expectations from the test should be considered before or after
`make` has been executed.

Including `executeMakeBeforeRun` in the TestScenario data class
prevents tests with otherwise identical setups from being run in the
same session.  Instead, pull out `executeMakeBeforeRun` into an outer
class containing the other parts of the scenario as a single
TestScenarioSetup data class.  Group the test definitions on this
TestScenarioSetup (as before), and verify the expectations of each
individual test definition before or after running make, as
appropriate.  (If there are no expectations for a given setup that
require running make, do not run make at all.)

This saves around 150 seconds for this test (split) target.

Bug: 333945762
Test: merged
Change-Id: Ie1929d36152470c78c5a8bd566728ba57b80cb52

GitOrigin-RevId: 702122ecf51dec4ca8a639c155b43c7e11430a3d