[android] FIX: com.android.tools.idea.res.ResourceClassGeneratorTest#testWithAars.testWithAars
This change fixes the test setup for IntelliJ monorepo, and should be skipped in AOSP.

com.android.tools.idea.res.ResourceClassGeneratorTest#testWithAars method requires `getTestAarRepositoryFromExplodedAar` setup method is located in a prebuilt module library of the`android.sdktools.testutils` module.
Original `getTestAarRepositoryFromExplodedAar` function references directory of the AOSP project which is not existing in IJ source code.
To fix the setup, we copied the implementation of `getTestAarRepositoryFromExplodedAar` function to `intellij.android.testFramework` module,
and we update implementation to reference test data in `community/android/android/testData/aar` directory.

GitOrigin-RevId: 98d27528d599cffa00e922c22f8356e06eb5c2ca